Create node.py
This node file runs on the device to communicate with hub and receive messages intended for it from hub and transmit messages to other users through hub. The Phonebook consists of usernames and their device and Node ids. the user_id here is the device id for the hub to send message to.